Revert "FEATURE: renames file to api.service"
This reverts commit d60541c36177ff2ea67ae69a336939202e8fa9f3.
