--- libs/core/data-access/src/lib/batch-update.interface.ts
@@ -1,6 +1,0 @@
-export interface BatchUpdate {
-  endpoint: string;
-  id: string;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  data: { [key: string]: any };
-}
