+++ libs/menu-matriarch/shared/data-access-api/src/index.ts
@@ -1,4 +1,4 @@
-export * from './lib/api.service';
+export * from './lib/batch.service';
 export * from './lib/dish-update.service';
 export * from './lib/endpoint.enum';
 export * from './lib/ingredient-type-update.service';
