+++ libs/menu-matriarch/shared/data-access-api/src/lib/dish-update.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { BatchUpdate } from '@atocha/core/data-access';
 
-import { ApiService, KeyToUpdate } from './api.service';
+import { ApiService, KeyToUpdate } from './batch.service';
 import { Endpoint } from './endpoint.enum';
 import {
   TallyChange,
